update spark_week_1.py and transfer of schemes and class date
